modified util.js

+++ util/file.js
@@ -1,9 +1,17 @@
 const fs = require("fs");
 
 exports.deleteFile = (filePath) => {
-  fs.unlink(filePath, (err) => {
-    if (err) {
-      throw err;
+  // Check if the file exists before attempting to delete it
+  fs.exists(filePath, (exists) => {
+    if (exists) {
+      fs.unlink(filePath, (err) => {
+        if (err) {
+          throw err; // Handle error if unlink fails
+        }
+        console.log('File deleted successfully');
+      });
+    } else {
+      console.log('File does not exist');
     }
   });
 };
